Switched to command handling system
