+++ commands/ls.py
@@ -1,0 +1,17 @@
+description = "This commands shows the list of files in cwd(current working directory)"
+
+from colorama import Fore
+import os
+# must have
+def run(args: str, ro_path):
+    path = ro_path
+    if len(args.split(" ")) > 1:
+        path = " ".join(args.split(" ")[1:])
+
+    for files in os.listdir(path):
+        print(Fore.CYAN + '-' + files)
+
+
+# must have
+def constructor():
+    return {"description": description, "exec": run}
