--- help_cmd.json
@@ -1,94 +1,0 @@
-{
-    "Commands": {
-        "cd": {
-            "definition": "This command is used to change directory.",
-            "usage_1": "If we type cd 'path'. It will go to the defined path by the user.",
-            "usage_2": "If the user types 'cd ../' It will go to the previous dir.",
-            "example_1": "cd Downloads' The directory will change to Downloads",
-            "example_2": "cd ../'The directory will go to the previous dir. C://(here)"
-        },
-
-
-        "echo": {
-            "definition": "This command will repeat the user input",
-            "usage": "echo 'message'. The output will be the message typed.",
-            "example": "echo Hello. The output will be Hello."
-        },
-
-        "mkdir": {
-            "definition": "This command creates new directory(folder).",
-            "usage": "mkdir 'name'. Here it will create a directory named the name given by the user",
-            "example": "mkdir Test. Here it will create a directory named Test"
-        },
-
-        "rmdir": {
-            "definition": "This command deletes directory(folder).",
-            "usage": "rmdir 'name od dir'. Here it will delete the directory which is selected by the user",
-            "example": "'rmdir Test'. Here it will delete the directory selected"
-        },
-
-
-        "create": {
-            "definition": "This command is used to create files in the selected dir",
-            "usage": "create 'name(with ext)'. Here it will create a file in the cwd(current working directory) name given in the user input",
-            "example": "'create file.txt'. Here it will create a file named file.txt on the cwd"
-        },
-
-        "rm": {
-            "definition": "This command remove the selected file in the cwd(current working directory)",
-            "usage": "rm 'name(withext)'. Here it will delete this file",
-            "example": "'rm file.txt'. Here it will delete this file.txt"
-        },
-
-        "search": {
-            "definition": "This command searches for an existing file in the cwd(current working directory)",
-            "usage": "search 'name(with ext)'. Here it shows the path of the searched file",
-            "example": "'search file.txt'. Here it shows the path of the file 'file.txt'"
-        },
-
-        "ls": {
-            "definition": "This commands shows the list of files in cwd(current working directory)",
-            "usage": "'ls'. It shows all the files and folders in the cwd(current workinf directory)"
-        },
-
-        "read": {
-            "definition": "This command show to information typed inside a file",
-            "usage": "read 'file name(with ext)'. It prints the charecters present in it.",
-            "example": "'read file.txt' If we type this command the Terminal prints the data present in it."
-        },
-
-        "pwd": {
-
-            "definition": "This command shows the pwd(present working directory)",
-            "usage": "'pwd'. The Terminal prints the directory your working in."
-        },
-
-        "clear": {
-
-            "definition": "This command clears the entire screen",
-            "usage": "'clear'. If we type this command the terminal clears everything on the screen"
-        },
-
-        "exec": {
-            "definition": "This command executes the command typed after it divided by a space",
-            "usage": "exec 'command'. It executes the command types after it.",
-            "example": "'exec echo'. The terminal executes echo function."
-        },
-
-        "stop/quit": {
-            "definition": "This command quits the terminal and closes",
-            "usage": "'stop'. The terminal will stop and close"
-        },
-
-        "whoami": {
-            "definition": "This command prints the 'user_id' in which the user is logged in",
-            "usage": "'whoami' This command prints the user using it."
-        },
-
-        "ping": {
-            "definition": "This command is used to check the ping of the the website mentioned",
-            "usage": "ping 'website'. It shows the ping of the website mentioned",
-            "message": "'ping google.com'. It shows the ping of google"
-        }
-    }
-}
